+++ share/objects/specials/malachite/01-malachite.lua
@@ -4,5 +4,33 @@ local globals = map.globals
 
 local id = require ('lib.gem.id')
 
--- Malachite
-local unit = objects [id (globals.Gem_Special__MALACHITE_1)]
+-- # Malachite
+do
+	local unit = objects [id (globals.Gem_Special__MALACHITE_1)]
+
+	-- ## Combat
+	do
+		-- Acquisition Range
+		unit.uacq.value = 800
+	end
+
+	-- ## Combat - Attack 1
+	do
+		-- Cooldown
+		unit.ua1c.value = 0.33
+
+		-- Range
+		unit.ua1r.value = 800
+	end
+end
+
+-- # Malachite Information
+do
+	local ability = objects.A03D
+
+	-- ## Data
+	do
+		-- Maximum Number of Targets
+		ability.Efk3.value = 9
+	end
+end