+++ CHANGELOG.md
@@ -55,6 +55,16 @@ a full list of changes to the project, please consult the commit log.
       up to `750` range.
 - Viper Slate:
     - Damage bonus gain reduced from `5%` to `3%` per level.
+- Malachite:
+    - Now attacks all targets in range.
+    - Attack range increased from `750` to `800`.
+    - Attack cooldown reduced from `0.5` to `0.33` seconds.
+- Vivid Malachite:
+    - Now attacks all targets in range.
+    - Attack range increased from `750` to `800`.
+    - Attack cooldown reduced from `0.5` to `0.33` seconds.
+- Might Malachite:
+    - Attack cooldown reduced from `0.55` to `0.33` seconds.
 - Dark Emerald / Enchanted Emerald:
     - Each additional Dark Emerald or Enchanted Emerald:
         - Causes attacks to bounce to an additional nearby enemy.